requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
google-generativeai>=0.3.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
rich>=13.0.0
typer>=0.9.0
pydantic>=2.0.0

